fixed RAND_MAX bug in info.hpp, 02 tests do run

+++ pbe-code/02-fast_slow_inheritance/info.hpp
@@ -44,7 +44,7 @@ namespace iot {
       
       item.temperature_kelvin += ((std::rand()/float(RAND_MAX)) - 1)*10;
       
-      item.power_consumption = ((std::rand()/RAND_MAX)-1.)*item.power_consumption;
+      item.power_consumption = ((std::rand()/double(RAND_MAX))-1.)*item.power_consumption;
     }
 
   }
